requirements.txt
fastapi
uvicorn
loguru

requests
pydantic

aiopg
aiopg[sa]
sqlalchemy-filters
